requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
future==0.17.1
idna==2.7
numpy==1.15.4
pandas==0.23.4
py-openaq==1.1.0
pycparser==2.19
python-dateutil==2.7.5
python-telegram-bot==11.1.0
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1
